Moved exports to config/exports, fixed codestyles

+++ scripts/install-composer.sh
@@ -1,19 +1,19 @@
-#!/bin/sh
+#!/bin/bash
 
-EXPECTED_CHECKSUM="$(php -r 'copy("https://composer.github.io/installer.sig", "php://stdout");')"
-php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
-ACTUAL_CHECKSUM="$(php -r "echo hash_file('sha384', 'composer-setup.php');")"
+if command -v php &> /dev/null; then
+  EXPECTED_CHECKSUM="$(php -r 'copy("https://composer.github.io/installer.sig", "php://stdout");')"
+  php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
+  ACTUAL_CHECKSUM="$(php -r "echo hash_file('sha384', 'composer-setup.php');")"
 
-if [ "$EXPECTED_CHECKSUM" != "$ACTUAL_CHECKSUM" ]
-then
-    >&2 echo 'ERROR: Invalid installer checksum'
+  if [ "$EXPECTED_CHECKSUM" != "$ACTUAL_CHECKSUM" ]; then
+    echo >&2 'ERROR: Invalid installer checksum'
     rm composer-setup.php
     exit 1
+  fi
+
+  php composer-setup.php --quiet
+  RESULT=$?
+  rm composer-setup.php
+  mv composer.phar ~/.local/bin/composer
+  exit $RESULT
 fi
-
-php composer-setup.php --quiet
-RESULT=$?
-rm composer-setup.php
-mv composer.phar ~/.local/bin/composer
-exit $RESULT
-
